+++ .gitignore
@@ -31,3 +31,4 @@ yarn-error.log*
 # vercel
 .vercel
 .old
+.npmcheckrc
