+++ components/layouts/nav-links.js
@@ -40,19 +40,20 @@ export default function NavLinks(props) {
       <HomeIcon className="w-6" style={{width: '50px'}} />
       Home
     </Link>
-    {props.path?
+    {props.path.length>=1?
       <Link
       key='Path'
       href='/viewPath'
       rel="noopener noreferrer"
       className={clsx(
-        'flex h-min grow items-center justify-center gap-2 rounded-md bg-gray-400 p-1 text-sm font-medium hover:bg-gray-300 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3',
+        'flex h-min items-center justify-center rounded-md bg-gray-400 p-1 text-sm font-medium hover:bg-gray-300 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3',
         {
           'bg-sky-100 text-blue-600': pathname === '/viewPath',
         },
       )}
+      style={{minWidth: '130px'}}
     >
-      <PresentationChartLineIcon className="w-6" style={{width: '50px'}} />
+      <PresentationChartLineIcon style={{minWidth: '50px'}} />
       View Path
     </Link>
       :<></>}
@@ -66,13 +67,13 @@ export default function NavLinks(props) {
             target={link.target?link.target:null}
             rel="noopener noreferrer"
             className={clsx(
-              'flex h-min grow items-center justify-center gap-2 rounded-md bg-gray-400 p-1 text-sm font-medium hover:bg-gray-300 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3',
+              'flex h-min grow items-center justify-center rounded-md bg-gray-400 p-1 text-sm font-medium hover:bg-gray-300 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3',
               {
                 'bg-sky-100 text-blue-600': pathname === link.href,
               },
             )}
           >
-            <LinkIcon className="w-6" style={{width: '50px'}} />
+            <LinkIcon style={{width: '50px'}} />
             {link.name}
           </Link>
         );
