+++ package.json
@@ -21,13 +21,13 @@
     "autoprefixer": "^10.4.19",
     "clsx": "^2.1.1",
     "colour-codes": "^1.0.8",
-    "expo": "~51.0.17",
-    "framer-motion": "^11.3.28",
-    "lucide-react": "^0.428.0",
+    "expo": "^51.0.30",
+    "framer-motion": "^11.3.29",
+    "lucide-react": "^0.429.0",
     "mongodb": "^6.8.0",
     "mongoose": "^8.5.2",
     "nativewind": "^2.0.11",
-    "next": "^14.2.4",
+    "next": "^14.2.6",
     "next-fonts": "^1.5.1",
     "postcss": "^8.4.39",
     "react": "^18.0.0",
@@ -38,9 +38,9 @@
     "tailwindcss": "^3.4.4"
   },
   "devDependencies": {
-    "@types/react": "18.3.3",
+    "@types/react": "^18.3.4",
     "autoprefixer": "^10.4.19",
-    "eslint-config-next": "14.2.4",
+    "eslint-config-next": "^14.2.6",
     "postcss": "^8.4.39",
     "tailwindcss": "^3.4.4"
   },
